backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tortoise-orm==0.20.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
jinja2==3.1.4
aiofiles==24.1.0
mangum==0.19.0